+++ XLaunch/LaunchesViewController.swift
@@ -1,0 +1,15 @@
+//
+//  ViewController.swift
+//  XLaunch
+//
+//  Created by Anastasia Lenina on 31.07.2023.
+//
+
+import UIKit
+
+class LaunchesViewController: UIViewController {
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    view.backgroundColor = .systemPink
+  }
+}
